+++ src/main/java/kitchenpos/order/dto/OrderRequest.java
@@ -24,4 +24,10 @@ public class OrderRequest {
         return orderLineItems.stream()
                 .map(OrderLineItemRequest::toEntity).collect(Collectors.toList());
     }
+
+    public List<Long> getMenuIds() {
+        return orderLineItems.stream()
+                .map(OrderLineItemRequest::getMenuId)
+                .collect(Collectors.toList());
+    }
 }
